adedd documentation and cleaned examples of modal and notify
