+++ src/jsNotify.php
@@ -14,6 +14,8 @@ class jsNotify implements jsExpressionable
     {
         if ($options && is_array($options)) {
             $this->setOptions($options);
+        } elseif (is_string($options)) {
+            $this->setOptions(['content'=>$options]);
         }
 
         if ($attachTo) {
